LCM refactor
