+++ cmd/root.go
@@ -241,9 +241,6 @@ func Initialize(isMigratedToLibrary, isBench bool) (err error) {
 
 }
 
-// hold a pointer to the global lifecycle controller so that commands could output messages and exit properly
-var glcm = common.GetLifecycleMgr()
-
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
