+++ cmd/syncEnumerator.go
@@ -250,6 +250,7 @@ func (cca *cookedSyncCmdArgs) initEnumerator(ctx context.Context) (enumerator *s
 		FileAttributes: common.FileTransferAttributes{
 			TrailingDot: cca.trailingDot,
 		},
+		JobErrorHandler: glcm,
 	}
 
 	var srcReauthTok *common.ScopedAuthenticator
