+++ cmd/zc_traverser_blob_versions.go
@@ -22,9 +22,10 @@ package cmd
 
 import (
 	"context"
+	"strings"
+
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blob"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/service"
-	"strings"
 
 	"github.com/Azure/azure-storage-azcopy/v10/common"
 )
@@ -66,7 +67,11 @@ func (t *blobVersionsTraverser) getBlobProperties(versionID string) (*blob.GetPr
 	if err != nil {
 		return nil, err
 	}
-	props, err := blobClient.GetProperties(t.ctx, &blob.GetPropertiesOptions{CPKInfo: t.cpkOptions.GetCPKInfo()})
+	cpkInfo, err := t.cpkOptions.GetCPKInfo()
+	if err != nil {
+		return nil, err
+	}
+	props, err := blobClient.GetProperties(t.ctx, &blob.GetPropertiesOptions{CPKInfo: cpkInfo})
 	return &props, err
 }
