+++ common/output.go
@@ -48,20 +48,6 @@ func (o OutputMessageType) String() string {
 	return enum.StringInt(o, reflect.TypeOf(o))
 }
 
-// defines the output and how it should be handled
-type outputMessage struct {
-	msgContent    string
-	msgType       OutputMessageType
-	exitCode      ExitCode      // only for when the application is meant to exit after printing (i.e. Error or Final)
-	inputChannel  chan<- string // support getting a response from the user
-	promptDetails PromptDetails
-}
-
-func (m outputMessage) shouldExitProcess() bool {
-	return m.msgType == EOutputMessageType.Error() ||
-		(m.msgType == EOutputMessageType.EndOfJob() && !(m.exitCode == EExitCode.NoExit()))
-}
-
 // used for output types that are not simple strings, such as progress and init
 // a given format(text,json) is passed in, and the appropriate string is returned
 type OutputBuilder func(OutputFormat) string
@@ -89,19 +75,6 @@ func GetJsonStringFromTemplate(template interface{}) string {
 	PanicIfErr(err)
 
 	return string(jsonOutput)
-}
-
-// defines the general output template when the format is set to json
-type JsonOutputTemplate struct {
-	TimeStamp      time.Time
-	MessageType    string
-	MessageContent string // a simple string for INFO and ERROR, a serialized JSON for INIT, PROGRESS, EXIT
-	PromptDetails  PromptDetails
-}
-
-func newJsonOutputTemplate(messageType OutputMessageType, messageContent string, promptDetails PromptDetails) *JsonOutputTemplate {
-	return &JsonOutputTemplate{TimeStamp: time.Now(), MessageType: messageType.String(),
-		MessageContent: messageContent, PromptDetails: promptDetails}
 }
 
 // Ideally this is just JobContext, but we probably shouldn't break the json output format
