+++ e2etest/declarativeScenario.go
@@ -631,10 +631,12 @@ func (s *scenario) validateContent() {
 				destName = addedDirAtDest + "/" + f.name
 			}
 			destName = fixSlashes(destName, s.fromTo.To())
+			cpkInfo, err := common.GetCpkInfo(s.p.cpkByValue)
+			s.a.AssertNoErr(err, "getting cpkInfo")
 			actualContent := s.state.dest.downloadContent(s.a, downloadContentOptions{
 				resourceRelPath: destName,
 				downloadBlobContentOptions: downloadBlobContentOptions{
-					cpkInfo:      common.GetCpkInfo(s.p.cpkByValue),
+					cpkInfo:      cpkInfo,
 					cpkScopeInfo: common.GetCpkScopeInfo(s.p.cpkByName),
 				},
 			})
@@ -708,11 +710,13 @@ func (s *scenario) validateSymlink(f *testObject, metadata map[string]*string) {
 			val, ok := metadata[common.POSIXSymlinkMeta]
 			c.Assert(ok, equals(), true)
 			c.Assert(*val, equals(), "true")
+			cpkInfo, err := common.GetCpkInfo(s.p.cpkByValue)
+			c.Assert(err, equals(), nil)
 
 			content := dest.downloadContent(c, downloadContentOptions{
 				resourceRelPath: fixSlashes(path.Join(addedDirAtDest, f.name), common.ELocation.Blob()),
 				downloadBlobContentOptions: downloadBlobContentOptions{
-					cpkInfo:      common.GetCpkInfo(s.p.cpkByValue),
+					cpkInfo:      cpkInfo,
 					cpkScopeInfo: common.GetCpkScopeInfo(s.p.cpkByName),
 				},
 			})
