+++ e2etest/newe2e_resource_managers_blob.go
@@ -522,7 +522,9 @@ func (b *BlobObjectResourceManager) CreateWithOptions(a Asserter, body ObjectCon
 			}
 		}
 
-		_, err := b.Container.InternalClient.NewBlockBlobClient(b.Path).UploadStream(ctx, body.Reader(), &blockblob.UploadStreamOptions{
+		cpkInfo, err := opts.CpkOptions.GetCPKInfo()
+		a.NoError("get CPKInfo", err)
+		_, err = b.Container.InternalClient.NewBlockBlobClient(b.Path).UploadStream(ctx, body.Reader(), &blockblob.UploadStreamOptions{
 			BlockSize:               blockSize,
 			Concurrency:             runtime.NumCPU(),
 			TransactionalValidation: blob.TransferValidationTypeComputeCRC64(),
@@ -530,7 +532,7 @@ func (b *BlobObjectResourceManager) CreateWithOptions(a Asserter, body ObjectCon
 			Metadata:                properties.Metadata,
 			AccessTier:              blobProps.BlockBlobAccessTier,
 			Tags:                    blobProps.Tags,
-			CPKInfo:                 opts.CpkOptions.GetCPKInfo(),
+			CPKInfo:                 cpkInfo,
 			CPKScopeInfo:            opts.CpkOptions.GetCPKScopeInfo(),
 		})
 		a.NoError("Block blob upload", err)
@@ -539,7 +541,9 @@ func (b *BlobObjectResourceManager) CreateWithOptions(a Asserter, body ObjectCon
 		client := b.Container.InternalClient.NewPageBlobClient(b.Path)
 		blockSize := DerefOrDefault(opts.BlockSize, common.DefaultPageBlobChunkSize)
 		size := body.Size()
-		_, err := client.Create(
+		cpkInfo, err := opts.CpkOptions.GetCPKInfo()
+		a.NoError("get CPKInfo", err)
+		_, err = client.Create(
 			ctx,
 			size,
 			&pageblob.CreateOptions{
@@ -547,7 +551,7 @@ func (b *BlobObjectResourceManager) CreateWithOptions(a Asserter, body ObjectCon
 				Metadata:     properties.Metadata,
 				Tier:         blobProps.PageBlobAccessTier,
 				HTTPHeaders:  properties.HTTPHeaders.ToBlob(),
-				CPKInfo:      opts.CpkOptions.GetCPKInfo(),
+				CPKInfo:      cpkInfo,
 				CPKScopeInfo: opts.CpkOptions.GetCPKScopeInfo(),
 			})
 		a.NoError("Page blob create", err)
@@ -567,13 +571,15 @@ func (b *BlobObjectResourceManager) CreateWithOptions(a Asserter, body ObjectCon
 			}
 			buf = buf[:n] // reduce buffer size for block
 
+			cpkInfo, err := opts.CpkOptions.GetCPKInfo()
+			a.NoError("get CPKInfo", err)
 			_, err = client.UploadPages(
 				ctx,
 				streaming.NopCloser(bytes.NewReader(buf)),
 				blob.HTTPRange{Offset: offset, Count: int64(n)},
 				&pageblob.UploadPagesOptions{
 					TransactionalValidation: blob.TransferValidationTypeComputeCRC64(),
-					CPKInfo:                 opts.CpkOptions.GetCPKInfo(),
+					CPKInfo:                 cpkInfo,
 					CPKScopeInfo:            opts.CpkOptions.GetCPKScopeInfo(),
 				})
 			a.NoError("Page blob upload", err)
@@ -593,9 +599,11 @@ func (b *BlobObjectResourceManager) CreateWithOptions(a Asserter, body ObjectCon
 
 		client := b.Container.InternalClient.NewAppendBlobClient(b.Path)
 
-		_, err := client.Create(ctx, &appendblob.CreateOptions{
+		cpkInfo, err := opts.CpkOptions.GetCPKInfo()
+		a.NoError("get CPKInfo", err)
+		_, err = client.Create(ctx, &appendblob.CreateOptions{
 			HTTPHeaders:  properties.HTTPHeaders.ToBlob(),
-			CPKInfo:      opts.CpkOptions.GetCPKInfo(),
+			CPKInfo:      cpkInfo,
 			CPKScopeInfo: opts.CpkOptions.GetCPKScopeInfo(),
 			Tags:         blobProps.Tags,
 			Metadata:     properties.Metadata,
@@ -617,13 +625,15 @@ func (b *BlobObjectResourceManager) CreateWithOptions(a Asserter, body ObjectCon
 			}
 			buf = buf[:n] // reduce buffer size for block
 
+			cpkInfo, err := opts.CpkOptions.GetCPKInfo()
+			a.NoError("get CPKInfo", err)
 			_, err = client.AppendBlock(ctx, streaming.NopCloser(bytes.NewReader(buf)), &appendblob.AppendBlockOptions{
 				TransactionalValidation: blob.TransferValidationTypeComputeCRC64(),
 				AppendPositionAccessConditions: &appendblob.AppendPositionAccessConditions{
 					AppendPosition: pointerTo(offset),
 					MaxSize:        pointerTo(offset + int64(n)),
 				},
-				CPKInfo:      opts.CpkOptions.GetCPKInfo(),
+				CPKInfo:      cpkInfo,
 				CPKScopeInfo: opts.CpkOptions.GetCPKScopeInfo(),
 			})
 			a.NoError("Append blob upload", err)
