+++ ste/mgr-JobPartTransferMgr.go
@@ -103,6 +103,7 @@ type IJobPartTransferMgr interface {
 	SuccessfulBytesTransferred() int64
 	TransferIndex() (partNum, transferIndex uint32)
 	RestartedTransfer() bool
+	GetJobErrorHandler() common.JobErrorHandler
 }
 
 // TransferInfo is a per path object that needs to be transferred
@@ -867,7 +868,7 @@ func (jptm *jobPartTransferMgr) failActiveTransfer(typ transferErrorCode, descri
 			}
 
 			if fileerror.HasCode(err, "ShareSizeLimitReached") {
-				common.GetLifecycleMgr().Error("Increase the file share quota and call Resume command.")
+				jptm.GetJobErrorHandler().Error("Increase the file share quota and call Resume command.")
 			}
 
 			// and use the normal cancelling mechanism so that we can exit in a clean and controlled way
@@ -1049,4 +1050,8 @@ func (jptm *jobPartTransferMgr) ShouldInferContentType() bool {
 
 func (jptm *jobPartTransferMgr) SuccessfulBytesTransferred() int64 {
 	return atomic.LoadInt64(&jptm.atomicSuccessfulBytes)
+}
+
+func (jptm *jobPartTransferMgr) GetJobErrorHandler() common.JobErrorHandler {
+	return jptm.jobPartMgr.GetJobErrorHandler()
 }
