+++ ste/xfer-deleteBlobFS.go
@@ -64,7 +64,7 @@ func doDeleteHNSResource(jptm IJobPartTransferMgr) {
 				if respErr.StatusCode == http.StatusForbidden {
 					errMsg := fmt.Sprintf("Authentication Failed. The SAS is not correct or expired or does not have the correct permission %s", err.Error())
 					jptm.Log(common.LogError, errMsg)
-					common.GetLifecycleMgr().Error(errMsg)
+					jptm.GetJobErrorHandler().Error(errMsg)
 				} else {
 					// in all other cases, make the transfer as failed
 					status = common.ETransferStatus.Failed()
